Indentation
This diagram shows that users need to log in before registration. If the information is valid, then login will be successful, or information is invalid, then login will be failed. Then the user can request their registration; if the information is valid, registration will be successful & it gives the message. If there is any failure to register, the user will get a return message of failed registration. After registration user’s information will be stored in the database. Also, users can request their 
update information; if the information is valid, then the update will be successful.